+++ snmp/tests/test_unit.py
@@ -15,8 +15,8 @@ from datadog_checks.base import ConfigurationError
 from datadog_checks.dev import temp_dir
 from datadog_checks.snmp import SnmpCheck
 from datadog_checks.snmp.config import InstanceConfig
+from datadog_checks.snmp.models import ObjectIdentity
 from datadog_checks.snmp.resolver import OIDTrie
-from datadog_checks.snmp.types import ObjectIdentity
 from datadog_checks.snmp.utils import oid_pattern_specificity, recursively_expand_base_profiles
 
 from . import common
